Minor bug fix and warning suppression.
The qsort() function requires that its predicate (in this case,
the group_info_cmp() function) use the __cdecl calling
convention. When a different default calling convention
is set globally for a project (e.g., /Gz or /Gv), this results
in a compiler failure, since the predicate has the wrong
calling convention. Adding an explicit annotation of
the __cdecl calling convention fixes this problem
with no side effects.

Elsewhere, an unused parameter was commented out
in the definition to avoid a warning.